v2/pyproject.toml
[project]
name = "schmith"
version = "0.1.0"
description = "Endpoint-focused C# DataObject generator for API specs."
requires-python = ">=3.11"

dependencies = [
    "pyyaml>=6.0.2",
    "ruamel.yaml>=0.18.6",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.0.0"]
all = ["anthropic>=0.40.0", "openai>=1.0.0"]

[project.scripts]
schmith = "schmith.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
